+++ packages/datafit/src/index.ts
@@ -1,9 +1,0 @@
-/**
- * @packageDocumentation
- * Simple curve-fitting algorithm
- * 
- * ![NPM Downloads](https://img.shields.io/npm/d18m/datafit)
- * ![NPM Last Update](https://img.shields.io/npm/last-update/datafit)
- */
-export * from './types';
-export * from './lib';