+++ packages/smath/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.0.0
+
+- Add `DataFit` submodule, can be imported via `import { DataFit } from 'smath'`
+- `import * as SMath from 'smath'` now changes to `import { SMath } from 'smath'`
+
 ## 1.16.0
 
 - Compatibility update for Typescript v6+